restore source-faithful schema parity across the application-data seam
Resolves all 8 code review findings for the Order-1 Application-Data /
Authoritative-Schema Seam checkpoint (1 CRITICAL, 6 MAJOR, 1 MINOR). The
authoritative reference throughout is the retired backend/app/db/models.py,
recovered from git history, per AAP G2 and review rule CR-SA3.

C1 CRITICAL - Schema type chain restored to 32-bit identifiers.
The identifier type had been taken from AAP 0.4.1.4's prose (JpaRepository
<Tweet, Long>) rather than the authoritative source column Column(Integer,
primary_key=True), producing a coherent but wrong Long/BIGINT chain; AiTool
was an outright regression from order-0. Converted the whole chain atomically
to Integer: Tweet, Response and AiTool id fields, accessors and equals locals;
all three JpaRepository type arguments; ResponseService (Integer.valueOf) and
TwitterService (Integer.parseInt); both mappers. IDENTITY generation and the
decimal-String wire boundary are unchanged, and an out-of-32-bit-range path
identifier is now reported as not-found rather than 400/500. The chain is
type-locked: reverting one entity to Long no longer compiles.

C2 MAJOR - One authorized delimited-list codec.
Withdrew the backslash escape grammar, which had created a second
representation while the Notion mirror kept the first and which silently
rewrote pre-existing column values containing a backslash. Restored the frozen
simple comma codec (AAP DL-024) and exposed it as encode/decode public
statics; NotionService now delegates both paths to it, so exactly one
implementation serves the column and the mirror.

C3 MAJOR - Persist-only per-key seeding.
saveAndFlush on an assigned String @Id routed through EntityManager.merge,
which could overwrite an operator row inserted concurrently. Seeding now
issues EntityManager.persist plus flush inside a PROPAGATION_REQUIRES_NEW
per-key transaction; a duplicate key is treated as already-exists and the
stored row is left exactly as it stands. merge is unreachable from the seed
path.

C4 MAJOR - Unbounded backlog query.
Replaced Slice<Tweet> findByResponsesIsEmptyOrderByIdAsc(Pageable) with
List<Tweet> findByResponsesIsEmpty(), restoring the source semantics of
Tweet.query.filter(Tweet.response == None).all().

C5 MAJOR - Test schema oracle narrowed and made source-derived.
Identifier and integer families are {Types.INTEGER} only, timestamps
{Types.TIMESTAMP} only, and the character family no longer admits fixed
CHAR/NCHAR. Added a three-dialect generated-DDL oracle (H2, PostgreSQL,
MySQL) asserting the required physical type for all 20 columns per vendor,
with negative rules forbidding bigint/int8/serial and any varchar in tweets,
responses or ai_tools.

C6 MAJOR - Rule 1 explainability lane.
Stripped rationale, alternatives and risk text from comments in the seven
cited files, leaving factual mapping, provenance and a DECISION_LOG pointer.
Rewrote in place DL-025, DL-069, DL-070, DL-138, DL-159, DL-164, DL-166 and
DL-179, allocated DL-182 for the backlog contract, and corrected two stale
citations. Realigned five TRACEABILITY_MATRIX rows.

C7 MAJOR - settings.key capacity.
Replaced the 255 bound with the measured maximum creatable and indexable as a
primary key on all three vendors AAP IR4 requires. Literal unboundedness is
provably impossible: Hibernate renders length=Integer.MAX_VALUE as clob/text/
longtext, H2 refuses a CLOB primary key, and MySQL 8.4 refuses every TEXT
family (ERROR 1170) and varchar(769) (ERROR 1071, 3072-byte utf8mb4 DYNAMIC
prefix limit). 768 characters creates and round-trips on H2, PostgreSQL 16 and
MySQL 8.4. settings.key now sits inside the same source-unbounded contract as
the other ten character columns.

C8 MINOR - Removed the redundant (JsonNode) cast in NotionServiceTest. The
tree is now warning-free under javac --release 21 -Xlint:all, where the
baseline carried one warning.

Verification: mvn clean verify BUILD SUCCESS, 423 tests (up from 408), zero
failures, zero warnings. Every new test proved non-vacuous by re-injecting the
original defect. The delivered schema was executed against real PostgreSQL 16
and MySQL 8.4; a runtime smoke test confirmed 32-bit identifier columns,
varchar(768) settings key, seeding, token issuance and the error envelopes,
and confirmed a restart leaves an operator-edited row's value and description
untouched.

No operations file, DTO, association mapping or security component changed.

+++ backend/src/main/java/com/codeskeptic/scanner/repository/AiToolRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * Spring Data JPA repository for the {@link AiTool} entity, which maps the {@code ai_tools} table.
  *
- * <p>The identifier type is {@link Long}, matching the {@code @Id} field of {@link AiTool} — see
+ * <p>The identifier type is {@link Integer}, matching the {@code @Id} field of {@link AiTool} — see
  * docs/DECISION_LOG.md DL-070. The {@code ai_tools} table declares no association. Both consumers
  * call the inherited surface:
  *
@@ -36,5 +36,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
  * @see AiTool
  */
 // Ported from backend/app/db/database.py:L10-13 (faithful port) — see docs/DECISION_LOG.md
-public interface AiToolRepository extends JpaRepository<AiTool, Long> {
+public interface AiToolRepository extends JpaRepository<AiTool, Integer> {
 }
